Indicator fix
Seems to be mostly successful, as far as I can tell with this initial
testing. Indicators increment exactly the correct amount. Added boolean
array for checking & moved indicator_change calll location to inside
create event of situation object
TODO: in the future, will need to find a way to reset the indicator
counter to false for random encounters